+++ invoice/views.py
@@ -2,17 +2,14 @@
 import io
 
 from django.http import FileResponse, HttpResponseRedirect
+from django.shortcuts import get_object_or_404
 from django.urls import reverse_lazy
 from django.views.generic import CreateView, UpdateView, DeleteView, ListView
 from django.views.generic import TemplateView
-from reportlab.lib import colors
-from reportlab.pdfgen import canvas
-from reportlab.platypus import Table
 
+from invoice import pdf_generator
 from invoice.forms import InvoiceItemForm, AddressForm, BankAccountForm, CustomerForm, VendorForm
 from invoice.models import Vendor, Customer, Invoice, InvoiceItem
-
-A4_WIDTH = 595
 
 
 class StartView(TemplateView):
@@ -124,72 +121,11 @@ class InvoiceListView(ListView):
     model = Invoice
 
 
-def get_y_position(y_print_start, y_step, y_steps):
-    """Calculate the y position of the next line on pdf canvas."""
-    return y_print_start - y_step * y_steps
-
-
 def pdf_invoice(request, invoice_id) -> FileResponse:
     """Generate an invoice as PDF file."""
+    invoice = get_object_or_404(Invoice, pk=invoice_id)
     buffer = io.BytesIO()
-    pdf_object = canvas.Canvas(buffer)
-    invoice = Invoice.objects.get(id=invoice_id)
-    y_print_start = 800
-    y_step = 20
-    y_steps = 0
-    pdf_object.drawString(100, get_y_position(y_print_start, y_step, y_steps), invoice.vendor.name)
-    y_steps += 1
-    if invoice.vendor.company_name:
-        pdf_object.drawString(100, get_y_position(y_print_start, y_step, y_steps),
-                              invoice.vendor.company_name)
-        y_steps += 1
-    pdf_object.drawString(100, get_y_position(y_print_start, y_step, y_steps),
-                          invoice.vendor.address.line_1)
-    y_steps += 1
-    if invoice.vendor.address.line_2:
-        pdf_object.drawString(100, get_y_position(y_print_start, y_step, y_steps),
-                              invoice.vendor.address.line_2)
-        y_steps += 1
-    if invoice.vendor.address.line_3:
-        pdf_object.drawString(100, get_y_position(y_print_start, y_step, y_steps),
-                              invoice.vendor.address.line_3)
-        y_steps += 1
-    pdf_object.drawString(100, get_y_position(y_print_start, y_step, y_steps),
-                          f'{invoice.vendor.address.postcode} {invoice.vendor.address.city}')
-    y_steps += 1
-    pdf_object.drawString(100, get_y_position(y_print_start, y_step, y_steps),
-                          invoice.vendor.address.country)
-    y_steps += 1
-    pdf_object.drawString(100, get_y_position(y_print_start, y_step, y_steps), 'Rechnung')
-    y_steps += 1
-
-    data = Invoice.objects.get(id=invoice_id).table_export
-    table = Table(
-        data=data,
-        style=[
-            ('GRID', (0, 0), (-1, -1), 0.5, colors.grey),
-            ('LEADING', (0, 0), (-1, 0), 10),
-            ('ALIGNMENT', (2, 1), (6, -1), "RIGHT"),
-        ]
-    )
-    _, table_height = table.wrapOn(pdf_object, 0, 0)
-
-    table_y_start = get_y_position(y_print_start, y_step, y_steps)
-    table.drawOn(pdf_object, x=100, y=table_y_start - table_height)
-
-    pdf_object.drawString(A4_WIDTH - 280, table_y_start - table_height - y_step,'Net Total:')
-    pdf_object.drawAlignedString(A4_WIDTH - 195, table_y_start - table_height - y_step,f'{invoice.net_total:.2f}')
-
-    pdf_object.drawString(A4_WIDTH - 280, table_y_start - table_height - y_step * 2, 'Total:')
-    pdf_object.drawAlignedString(A4_WIDTH - 195, table_y_start - table_height - y_step * 2, f'{invoice.total:.2f}')
-
-    if invoice.vendor.tax_id:
-        pdf_object.drawString(100, 100, f'Tax ID: {invoice.vendor.tax_id}')
-    if invoice.vendor.bank_account:
-        pdf_object.drawString(100, 80, f'IBAN: {invoice.vendor.bank_account.iban}')
-        pdf_object.drawString(100, 60, f'BIC: {invoice.vendor.bank_account.bic}')
-    pdf_object.showPage()
-    pdf_object.save()
+    _ = pdf_generator.gen_invoice_pdf(invoice, buffer)
     buffer.seek(0)
     return FileResponse(buffer, as_attachment=False, filename='invoice.pdf')
